requirements.txt
flask==3.1.0
flask-cors==5.0.1
openai==1.68.0
transformers==4.48.3
torch==2.6.0
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.3
python-dotenv==1.0.1
gunicorn==23.0.0
jinja2==3.1.5
pyyaml==6.0.2
